backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0.post1
SQLAlchemy==2.0.23
psycopg2-binary
alembic==1.13.0
python-dotenv==1.0.0
